% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnrexample1.R
\docType{data}
\name{cnrexample1}
\alias{cnrexample1}
\title{Simulated dataset with random uniform responses.}
\format{
A data frame with 300 rows and 18 columns.
}
\usage{
cnrexample1
}
\description{
A simulated dataset with uniform random responses.
The first 75 rows are fully attentive.
The remaining 225 rows have all item responses independently drawn from a 
uniform distribution.
}
\keyword{datasets}
