% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Classification performance metrics}
\usage{
metrics(y, yhat)
}
\arguments{
\item{y}{vector of true class labels. 
Can be numeric (0 vs 1) or character ("nonrandom vs random") format.}

\item{yhat}{vector of predicted class labels. 
Can be numeric (0 vs 1) or character ("spare" vs "flag") format.}
}
\value{
List of two components. 
First component is a confusion matrix, 
and second component is a vector of accuracy, specificity, sensitivity, 
positive predictive value, negative predictive value, 
and flag rate.
}
\description{
Given true class labels and predicted class labels, compute metrics like accuracy.
}
\examples{
set.seed(47)
y = rep(0:1, times=6) # true class labels
yhat = rep(0:1, each=6) # predicted class labels
metrics(y, yhat) # get confusion matrix, accuracy, etc
}
