% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcnrbinom.R
\name{rcnrbinom}
\alias{rcnrbinom}
\title{Generate content-nonresponsive data with binomial distribution}
\usage{
rcnrbinom(n, pointscales)
}
\arguments{
\item{n}{the sample size.}

\item{pointscales}{vector of integers indicating how many Likert-type response categories there are for each item. 
Length must match the number of columns in \code{data}.}
}
\value{
The sample matrix.
}
\description{
Generate a Likert-type sample where items are independent and 
follow a fair-coin binomial distribution, 
provided the number of response categories for each item.
Likert-type categories are from 1 (not 0) to K where K is the highest category.
}
\examples{
set.seed(47)
rcnrbinom(30, c(4, 4, 4, 5, 5, 5, 6, 6, 6)) # 30 participants, 9 items
}
\seealso{
\code{\link[stats]{dbinom}}, \code{\link{sample}}, \code{\link{rcnrunif}}
}
