% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcnrstring.R
\name{rcnrstring}
\alias{rcnrstring}
\title{Generate content-nonresponsive data with a particular category having specified probability}
\usage{
rcnrstring(n, pointscales, prob = 0.75, whichcat = NULL, allsame = TRUE)
}
\arguments{
\item{n}{the sample size.}

\item{pointscales}{vector of integers indicating how many Likert-type response categories there are for each item. 
Length must match the number of columns in \code{data}.}

\item{prob}{Scalar or vector indicating probability of selecting desired
categories specified in \code{whichcat}.}

\item{whichcat}{Scalar or vector indicating which category should have
assigned probability.}

\item{allsame}{Boolean value indicating whether each participant should
have the same response process.}
}
\value{
The sample matrix.
}
\description{
Generate a Likert-type sample where items are independent and 
a single category has a specified (possibly high) probability of endorsement.
}
\details{
The goal of this function is to mimic the case where a participant is
nonresponsive and selects (mostly) the same category throughout the survey.
Multiple nonresponsive participants may have the exact same behavior, or may
differ.

If \code{whichcat} is \code{NULL}, a random category is chosen with uniform
probability for each subscale with the same number of categories. For
example, if \code{pointscales} contains a vector of 5's, a value inclusive
of 1 through 5 is chosen with equal probability.

If \code{whichcat} is a scalar value, this single category is used across
all items. This choice requires that this scalar value is compatible with
\code{pointscales}. It must be a category that could be observed for all
items.

If \code{whichcat} is a vector, then is should have the same length as
\code{pointscales} (i.e., the number of items) and each entry should be a
valid category for each item.

For each item, whatever category is chosen using the above scheme is assigned
probability equal to \code{prob}. The remaining categories have equal
probability out of whatever is left, 1-\code{prob}.

\code{allsame} toggles whether content nonresponsive participants all have
the same response process or not. If \code{TRUE}, all will have same process.
If \code{FALSE}, then \code{whichcat} should be specified as \code{NULL} so
that different participants have a different favorite category that is
randomly generated. Note that a wrapper or multiple calls with generating
subsets of participants from this function could effectively generate
something similar if more fine-grained control is desired.

Likert-type categories are from 1 (not 0) to K where K is the highest category.
}
\examples{
set.seed(47)

# 50 participants, 10 5 category items, random fav cat, chosen at prob = .75
rcnrstring(50, rep(5, 10), prob=.75, whichcat=NULL)

# 50 participants, 10 5 category items, category 2 chosen at prob = .75
rcnrstring(50, rep(5, 10), prob=.75, whichcat=2)

# 50 participants, each with different preference at .80 prob
rcnrstring(50, rep(5, 10), prob=.8, whichcat=NULL, allsame=FALSE)

}
\seealso{
\code{\link{sample}}, \code{\link{rcnrunif}}, \code{\link{rcnrbinom}},
 \code{\link{rcnrreverse}}, \code{\link{rcnrpat}}
}
