% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcnrunif.R
\name{rcnrunif}
\alias{rcnrunif}
\title{Generate content-nonresponsive data with uniform distribution}
\usage{
rcnrunif(n, pointscales)
}
\arguments{
\item{n}{the sample size.}

\item{pointscales}{vector of integers indicating how many Likert-type response categories there are for each item. 
Length must match the number of columns in \code{data}.}
}
\value{
The sample matrix.
}
\description{
Generate a Likert-type sample where items are independent and uniformly 
distributed, 
provided the number of response categories for each item.
Likert-type categories are from 1 (not 0) to K where K is the highest category.
}
\examples{
set.seed(47)
rcnrunif(30, c(4, 4, 4, 5, 5, 5, 6, 6, 6)) # 30 participants, 9 items
}
\seealso{
\code{\link{sample}}, \code{\link{rcnrbinom}}
}
