% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplerows.R
\name{samplerows}
\alias{samplerows}
\title{Sample rows from a dataset}
\usage{
samplerows(n, data)
}
\arguments{
\item{n}{the number of rows to sample.}

\item{data}{the dataset to sample rows from.}
}
\value{
The sample. Has the same class as \code{data}.
}
\description{
Sample rows with replacement from a dataset.
}
\examples{
set.seed(47)
samplerows(10, datasets::iris)
}
\seealso{
\code{\link{sample}}
}
