% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnrdetect.R
\name{cnrdetect}
\alias{cnrdetect}
\title{Full content-nonresponsivity detection}
\usage{
cnrdetect(
  data,
  pointscales,
  numperms = 1000,
  feat_funs = c(mahal, ptcossim),
  feat_idvals = c(0, +1)
)
}
\arguments{
\item{data}{Likert-type data (matrix or dataframe).}

\item{pointscales}{vector of integers indicating how many Likert-type response 
categories there are for each item. 
Length must match the number of columns in \code{data}.}

\item{numperms}{an integer indicating how many permutations to produce per respondent.}

\item{feat_funs}{list of nonresponsivity feature functions. Each function must 
return a numeric vector.}

\item{feat_idvals}{vector of ideal values, in the same order as \code{feat_funs}.}
}
\value{
Vector of p-values. Smaller p-values are less suspicious.
}
\description{
Compute p-values for each respondent.
If the number of response options is the same for all items, 
the null hypothesis is that the response vector is exchangeable.
If the number of response options is not all the same, 
the null hypothesis is that: 
each subset of items that have the same number of response options is exchangeable; 
and subsets of items having different numbers of response options are independent.
}
\details{
Likert-type response data must be 1 to K where K is the highest response category. 
Missing values must be \code{NA}.

For the row tested, missing items are excluded from the test.
When no more than one item is nonmissing, p-value defaults to \code{NA}.
For the anchor sample, cells that have not been excluded but are missing are 
imputed with the midrange (e.g. "3" on a scale of 1 to 5).

When there are issues computing feature-space covariances, 
p-value defaults to \code{NA}.
}
\examples{
set.seed(47)
cnrdetect(cnrexample1, pointscales=rep(5, times=ncol(cnrexample1)), 
numperms=1000) # p values
}
