% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahal.R
\name{mahal}
\alias{mahal}
\title{Mahalanobis distance}
\usage{
mahal(x, ref)
}
\arguments{
\item{x}{data (matrix or dataframe) of interest.}

\item{ref}{reference sample (matrix or dataframe). Must have same number of columns as \code{x}.}
}
\value{
Vector of Mahalanobis distances.
}
\description{
Calculates the Mahalanobis distance with respect to some reference sample.
}
\seealso{
\code{\link[stats]{mahalanobis}}
}
