% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptcor.R
\name{ptcor}
\alias{ptcor}
\title{Person-total correlation}
\usage{
ptcor(x, ref)
}
\arguments{
\item{x}{data (matrix or dataframe) of interest.}

\item{ref}{reference sample (matrix or dataframe). Must have same number of columns as \code{x}.}
}
\value{
Vector of person-total correlations. Undefined values default to -1.
}
\description{
Calculates the person-total correlation with respect to some reference sample.
}
\seealso{
\code{\link[stats]{cor}}
}
